complete comment module

+++ app/Http/Controllers/BlogControler.php
@@ -6,7 +6,8 @@ use App\Models\Blog;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use App\Http\Requests\BlogRequest;
-
+use App\Http\Requests\UpdateBlogRequest;
+use Illuminate\Support\Facades\Storage;
 class BlogControler extends Controller
 {
     /**
@@ -38,10 +39,15 @@ class BlogControler extends Controller
       $data = $request->validated();
       //store image in public/images
       $image = $request->image;
+      //change the name of the image
       $NewImageName = time() . '.' . $image->getClientOriginalName();
+      //store the new image
       $image->storeAs('public/blogs', $NewImageName);
+      //update the data
       $data['image'] = $NewImageName;
+      //set the user id of the blog
       $data['user_id'] = Auth::user()->id;
+      //create the blog
       Blog::create($data);
       return redirect()->back()->with('blog_created', 'Blog created successfully');
         
@@ -61,25 +67,58 @@ class BlogControler extends Controller
      */
     public function edit(Blog $blog)
     {
-        //
+        if(Auth::check() && Auth::user()->id == $blog->user_id){
+            return view('theme.blogs.edit', compact('blog'));
+        }
+        abort(403, 'Unauthorized action.');
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, Blog $blog)
+    public function update( UpdateBlogRequest $request, Blog $blog)
     {
-        //
-    }
+        if(Auth::check() && Auth::user()->id == $blog->user_id){
+        
+        $data = $request->validated();
+        if($request->hasFile('image')){
+            //delete the old image
+            Storage::delete('public/blogs/' . $blog->image);
+            //get the old image
+            $image = $request->image;
+            //change the name of the image
+            $NewImageName = time() . '.' . $image->getClientOriginalName();
+            //store the new image
+            $image->storeAs('public/blogs', $NewImageName);
+            //update the data   
+            $data['image'] = $NewImageName;
+        }
+        //update the blog
+        $blog->update($data);
+        //redirect to the my blogs page
+        return redirect()->back()->with('blog_updated', 'Blog updated successfully');
+        }
+        abort(403, 'Unauthorized action.');
+    }       
 
     /**
      * Remove the specified resource from storage.
      */
     public function destroy(Blog $blog)
     {
-        $blog->delete();
-        return redirect()->back()->with('blog_deleted', 'Blog deleted successfully');
+        if(Auth::check() && Auth::user()->id == $blog->user_id){
+            //delete the old image
+            Storage::delete('public/blogs/' . $blog->image);
+            //delete the blog
+            $blog->delete();
+            //redirect to the my blogs page
+            return redirect()->back()->with('blog_deleted', 'Blog deleted successfully');
+        }
+        abort(403, 'Unauthorized action.');
     }
+
+
+
      /**
      * Display all the blogs of the user.
      */
